simple map works; tests separated; de-aliased
